Track the installed borg version to fix CI
When borg is not installed, we do not have a fact with the current
version and assume it is 1.2.0 to enable post-1.2.0 configuration.

On legacy systems, we install an older version of borg, so on next run,
the fact present the actual version, and the config is updated to match
the expected settings.

With this commit, we ensure the install is idempotent as expected by our
CI.